Update
Visual Improvements
- Replaced tile visuals
- Fixed transitioning between scenes/stories
New feature:
- Stunball (NEW): When hit, character will not be able to move for a short period of time.

Story:
- Added new story (DIO)

Backend:
- Fixed logic in some areas
- Added new attribute "name" to obstacles

Other:
Added new font

TODO List:
- Make time an object in SettingsSingleton allowing for universal use
- Create intro
- Remake game main menu music

+++ src/Map.java
@@ -11,8 +11,8 @@ public abstract class Map {
     Image map;
     Point pos;
     ArrayList<Tile> tiles;
-    int height;
-    int currentHeight;
+    double height;
+    double currentHeight;
     boolean jotaroAbility = false;
 
     public void draw() {
@@ -42,11 +42,11 @@ public abstract class Map {
         ArrayList<Tile> visibleTiles = new ArrayList<>();
         return visibleTiles;
     }
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
 
-    public int getCurrentHeight() {
+    public double getCurrentHeight() {
         return currentHeight;
     }
